typings files restructure
