+++ src/components/Buttons/typings.ts
@@ -1,0 +1,9 @@
+import {
+  TAnyFunction,
+} from '../../typings/definitions';
+
+export interface IButtonProps {
+  isHorizontal: boolean;
+  previousSlide: TAnyFunction;
+  nextSlide: TAnyFunction;
+}
