+++ src/components/ButtonsNav/typings.ts
@@ -1,0 +1,11 @@
+import {
+  IMenuNavProps,
+} from '../../typings/definitions';
+
+/**
+ * `ButtonsNav` component props.
+ */
+export interface IButtonsNavProps extends IMenuNavProps {
+  backgroundColor: string;
+  alignItems: string;
+}
