+++ src/components/MenuNav/typings.ts
@@ -1,0 +1,16 @@
+import {
+  INavProps,
+} from '../../typings/definitions';
+
+/**
+ * `MenuNav` component props.
+ */
+export interface IMenuNavProps extends INavProps {
+  navDescriptions: string[];
+  justifyContent: string;
+  sliderWidth: number;
+  mobileThreshold: number;
+  isNullAfterThreshold: boolean;
+  extraButton: React.ReactElement | React.Component;
+  isExtraButtonRight: boolean;
+}
