+++ src/components/SideNav/typings.ts
@@ -1,0 +1,12 @@
+import {
+  INavProps,
+} from '../../typings/definitions';
+
+/**
+ * `SideNav` component props.
+ */
+export interface ISideNavProps extends INavProps {
+  right: string;
+  left: string;
+  isPositionedRight: boolean;
+}
