+++ src/components/Slide/Mask/typings.ts
@@ -1,0 +1,9 @@
+import {
+  IBackgroundProps,
+} from '../../../typings/definitions';
+
+export interface IMaskProps {
+  isActive: boolean;
+  isDoneSliding: boolean;
+  background: IBackgroundProps;
+}
