+++ src/dependencies/isMobile.ts
@@ -1,11 +1,11 @@
 export const isMobile = (): boolean => {
   // Detect if running on the browser:
-  const isBrowser = typeof window !== 'undefined'
+  const isBrowser = typeof window !== 'undefined';
   if (isBrowser) {
     if (navigator && /Mobi|Android/i.test(navigator.userAgent)) {
-      return true
+      return true;
     }
-    return false
+    return false;
   }
-  return false
-}
+  return false;
+};
